feat: 모바일, pc 헤더 분리 및 분기점 hook 추가
