+++ src/components/common/MobileHeader.tsx
@@ -1,0 +1,13 @@
+import S from './MobileHeader.module.css'
+
+function MobileHeader() {
+  return (
+    <div>
+      <link
+        rel="stylesheet"
+        href=""
+      />
+    </div>
+  )
+}
+export default MobileHeader
